+++ src/marknotes/plugins/content/html/search.php
@@ -1,0 +1,65 @@
+<?php
+
+namespace MarkNotes\Plugins\Content\HTML;
+
+defined('_MARKNOTES') or die('No direct access allowed');
+
+class Search
+{
+
+    /**
+     * Provide additionnal javascript
+     */
+    public static function addJS(&$js = null)
+    {
+        $aeFunctions = \MarkNotes\Functions::getInstance();
+        $aeSettings = \MarkNotes\Settings::getInstance();
+
+        $root = rtrim($aeFunctions->getCurrentURL(true, false), '/');
+
+        if ($aeSettings->getDebugMode()) {
+            $js .= "\n<!-- Lines below are added by ".__FILE__."-->";
+        }
+
+        $js .= "\n<script type=\"text/javascript\" src=\"".$root."/marknotes/plugins/content/html/search/search.js\"></script>\n";
+
+        $js .=
+            "<script type=\"text/javascript\">\n".
+            "marknotes.message.search_no_result='".$aeSettings->getText('search_no_result', 'Sorry, the search is not successfull', true)."';\n".
+            "marknotes.settings.search_max_width=".SEARCH_MAX_LENGTH.";\n".
+            "</script>\n";
+
+        if ($aeSettings->getDebugMode()) {
+            $js .= "<!-- End for ".__FILE__."-->";
+        }
+
+        return true;
+    }
+
+    /**
+     * Provide additionnal stylesheets
+     */
+    public static function addCSS(&$css = null)
+    {
+        $aeFunctions = \MarkNotes\Functions::getInstance();
+        $aeSettings = \MarkNotes\Settings::getInstance();
+
+        $root = rtrim($aeFunctions->getCurrentURL(true, false), '/');
+
+        $css .=
+            "<link media=\"screen\" rel=\"stylesheet\" type=\"text/css\" href=\"".$root."/marknotes/plugins/content/html/search/search.css\">\n";
+
+        return true;
+    }
+
+    /**
+     * Attach the function and responds to events
+     */
+    public function bind()
+    {
+        $aeEvents = \MarkNotes\Events::getInstance();
+        $aeEvents->bind('render.js', __CLASS__.'::addJS');
+        $aeEvents->bind('render.css', __CLASS__.'::addCSS');
+        return true;
+    }
+}